Change the strategy pattern structure for the Assignment Conclude operations
