+++ src/main/java/com/ays/assignment/service/impl/handler/AssignmentCompleteHandler.java
@@ -5,9 +5,11 @@ import com.ays.assignment.model.entity.AssignmentEntity;
 import com.ays.assignment.model.enums.AssignmentHandlerType;
 import com.ays.assignment.model.enums.AssignmentStatus;
 import com.ays.assignment.repository.AssignmentRepository;
+import com.ays.assignment.util.exception.AysAssignmentNotExistByUserIdAndStatusException;
 import com.ays.auth.model.AysIdentity;
 import com.ays.user.model.entity.UserEntity;
 import com.ays.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 /**
@@ -15,14 +17,12 @@ import org.springframework.stereotype.Component;
  * This class is responsible for handling the {@link AssignmentEntity} with {@link AssignmentStatus#IN_PROGRESS} status.
  */
 @Component
-class AssignmentCompleteHandler extends AssignmentAbstractHandler {
+@RequiredArgsConstructor
+class AssignmentCompleteHandler implements AssignmentHandler {
 
     private final UserRepository userRepository;
-
-    public AssignmentCompleteHandler(AssignmentRepository assignmentRepository, UserRepository userRepository, AysIdentity identity) {
-        super(assignmentRepository, identity);
-        this.userRepository = userRepository;
-    }
+    private final AssignmentRepository assignmentRepository;
+    private final AysIdentity identity;
 
     @Override
     public AssignmentHandlerType type() {
@@ -31,18 +31,21 @@ class AssignmentCompleteHandler extends AssignmentAbstractHandler {
 
 
     @Override
-    protected AssignmentEntity handle(AssignmentEntity assignment) {
-        UserEntity user = assignment.getUser();
+    public void handle() {
+        AssignmentEntity assignmentEntity = this.findAssignmentEntity();
+        UserEntity user = assignmentEntity.getUser();
         user.ready();
         userRepository.save(user);
 
-        assignment.done();
-        return assignment;
+        assignmentEntity.done();
+        assignmentRepository.save(assignmentEntity);
     }
 
-    @Override
-    protected AssignmentStatus getAssignmentSearchStatus() {
-        return AssignmentStatus.IN_PROGRESS;
+    private AssignmentEntity findAssignmentEntity() {
+        String userId = identity.getUserId();
+        AssignmentStatus assignmentStatus = AssignmentStatus.IN_PROGRESS;
+        return assignmentRepository
+                .findByUserIdAndStatus(userId, assignmentStatus)
+                .orElseThrow(() -> new AysAssignmentNotExistByUserIdAndStatusException(userId, assignmentStatus));
     }
-
 }
