+++ src/main/java/com/ays/assignment/service/impl/handler/AssignmentStartHandler.java
@@ -5,9 +5,11 @@ import com.ays.assignment.model.entity.AssignmentEntity;
 import com.ays.assignment.model.enums.AssignmentHandlerType;
 import com.ays.assignment.model.enums.AssignmentStatus;
 import com.ays.assignment.repository.AssignmentRepository;
+import com.ays.assignment.util.exception.AysAssignmentNotExistByUserIdAndStatusException;
 import com.ays.auth.model.AysIdentity;
 import com.ays.user.model.entity.UserEntity;
 import com.ays.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 /**
@@ -15,34 +17,34 @@ import org.springframework.stereotype.Component;
  * This class is responsible for handling the {@link AssignmentEntity} with {@link AssignmentStatus#ASSIGNED} status.
  */
 @Component
-class AssignmentStartHandler extends AssignmentAbstractHandler {
+@RequiredArgsConstructor
+class AssignmentStartHandler implements AssignmentHandler {
 
     private final UserRepository userRepository;
-
-    public AssignmentStartHandler(AssignmentRepository assignmentRepository, UserRepository userRepository, AysIdentity identity) {
-        super(assignmentRepository, identity);
-        this.userRepository = userRepository;
-    }
+    private final AssignmentRepository assignmentRepository;
+    private final AysIdentity identity;
 
     @Override
     public AssignmentHandlerType type() {
         return AssignmentHandlerType.START;
     }
 
-
     @Override
-    protected AssignmentEntity handle(AssignmentEntity assignment) {
-        UserEntity user = assignment.getUser();
-        user.onRoad();
-        userRepository.save(user);
+    public void handle() {
+        AssignmentEntity assignmentEntity = this.findAssignmentEntity();
+        UserEntity userEntity = assignmentEntity.getUser();
+        userEntity.onRoad();
+        userRepository.save(userEntity);
 
-        assignment.start();
-        return assignment;
+        assignmentEntity.start();
+        assignmentRepository.save(assignmentEntity);
     }
 
-    @Override
-    protected AssignmentStatus getAssignmentSearchStatus() {
-        return AssignmentStatus.ASSIGNED;
+    private AssignmentEntity findAssignmentEntity() {
+        String userId = identity.getUserId();
+        AssignmentStatus assignmentStatus = AssignmentStatus.ASSIGNED;
+        return assignmentRepository
+                .findByUserIdAndStatus(userId, assignmentStatus)
+                .orElseThrow(() -> new AysAssignmentNotExistByUserIdAndStatusException(userId, assignmentStatus));
     }
-
 }
